+++ bigml/bigmlconnection.py
@@ -300,6 +300,7 @@ class BigMLConnection(object):
                                       location, resource, error)
             else:
                 try:
+                    print body
                     response = requests.post(url + self.auth,
                                              headers=SEND_JSON,
                                              data=body, verify=verify)
@@ -311,6 +312,7 @@ class BigMLConnection(object):
                     return maybe_save(resource_id, self.storage, code,
                                       location, resource, error)
             try:
+                print response.content
                 code = response.status_code
                 if code in [HTTP_CREATED, HTTP_OK]:
                     if 'location' in response.headers: